+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Using cron scheduler library from Rob Figueiredo
+
 ## [3.1.0] - 2017-06-21
 ### Added
 - Report e-mail in HTML format