+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Using cron scheduler library from Rob Figueiredo
 - Deploy scripts now configure themselves as daemons
+- Updated project dependencies
 
 ## [3.1.0] - 2017-06-21
 ### Added